requirements.txt
# GitHub App and API dependencies
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0
